// src/infrastructure/grpc/generated/course/types/enrollment.ts
export interface EnrollmentData {
  id: string;
  userId: string;
  courseId: string;
  status: string;
  progress: number;
  enrolledAt: string;
  completedAt: string;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  course: EnrollmentCourseData | undefined;
}

export interface EnrollmentCourseData {
  id: string;
  title: string;
  rating: number;
  thumbnail: string;
  category: string;
  level: string;
  lessonsCount: number;
  instructor: User | undefined;
}